Add proper failure messages and change IO method.
Using tagof() we can determine the type of values passed into SMIS and so
output proper failure messages when a test doesn't pass. These messages
format the 'expected' and 'got' values according to their type.

Rather than storing everything in arrays and aggregating the output at the
end, the plugin will now write the results to a file and read it back when
the test output is requested. Eventually it would be nice to have a way to
specify the filename, and have a way to download it at the end.